requirements.txt
python-dotenv==1.0.0
pandas==2.1.4
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
